requirements.txt
flask
flask-cors
flask-jsonpify
pyjwt
cryptography
boto
raven==6.0.0
